Small code refactoring improvements
