+++ winmdj-core/src/main/kotlin/com/github/danielchemko/winmdj/core/autoobject/stubs/StubTypeDefinitionImpl.kt
@@ -3,7 +3,7 @@ package com.github.danielchemko.winmdj.core.autoobject.stubs
 
 import com.github.danielchemko.winmdj.core.MdObjectMapper
 import com.github.danielchemko.winmdj.core.autoobject.BaseWinMdStub
-import com.github.danielchemko.winmdj.core.autoobject.model.CLRMetadataType
+import com.github.danielchemko.winmdj.core.mdspec.CLRMetadataType
 import com.github.danielchemko.winmdj.core.mdspec.TypeDefinition
 import com.github.danielchemko.winmdj.core.mdspec.WinMdStub
 import com.github.danielchemko.winmdj.parser.WinMdNavigator
